Revert "Merged changes into remote master"
This reverts commit 5d912916b02d4f873f7759c87c284d8cb47edbc2, reversing
changes made to ec800570404eb40a407e4222aae18c687c4a6b3e.

+++ include/EntityData.h
@@ -16,7 +16,7 @@ struct EntityData {
 	int minerals;
 	int vespene;
 	int buildTime;
-	float supplyCost; //can be 0.5 for zerglings
+	double supplyCost; //can be 0.5 for zerglings
 	int supplyProvided;
 	int startEnergy;
 	int maxEnergy;
